# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "K-Admin 后台管理系统 API 文档",
        "title": "K-Admin System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tools/gen/generate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据配置生成后端和前端代码，并写入文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Generator"
                ],
                "summary": "生成代码",
                "parameters": [
                    {
                        "description": "生成配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tools.GenerateConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回生成的文件路径列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/tools/gen/metadata/{tableName}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定表的元数据信息，包括列名、类型、约束等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Generator"
                ],
                "summary": "获取表元数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "表名",
                        "name": "tableName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/tools.TableMetadata"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/tools/gen/preview": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据配置生成代码预览，不写入文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Generator"
                ],
                "summary": "预览生成的代码",
                "parameters": [
                    {
                        "description": "生成配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tools.GenerateConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功，返回文件路径和内容的映射",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/tools/gen/table": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据字段定义创建新的数据库表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code Generator"
                ],
                "summary": "创建数据库表",
                "parameters": [
                    {
                        "description": "创建表请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "失败",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "tools.CodeGenColumnInfo": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "default": {
                    "type": "string"
                },
                "extra": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nullable": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "tools.FieldConfig": {
            "type": "object",
            "properties": {
                "column_name": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "form_type": {
                    "type": "string"
                },
                "gorm_tag": {
                    "type": "string"
                },
                "is_primary_key": {
                    "type": "boolean"
                },
                "json_tag": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "nullable": {
                    "type": "boolean"
                },
                "searchable": {
                    "type": "boolean"
                },
                "ts_type": {
                    "type": "string"
                }
            }
        },
        "tools.GenerateConfig": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tools.FieldConfig"
                    }
                },
                "frontend_path": {
                    "type": "string"
                },
                "module_path": {
                    "type": "string"
                },
                "options": {
                    "$ref": "#/definitions/tools.GenerateOptions"
                },
                "package_name": {
                    "type": "string"
                },
                "router_path": {
                    "type": "string"
                },
                "struct_name": {
                    "type": "string"
                },
                "table_comment": {
                    "type": "string"
                },
                "table_name": {
                    "type": "string"
                }
            }
        },
        "tools.GenerateOptions": {
            "type": "object",
            "properties": {
                "generate_api": {
                    "type": "boolean"
                },
                "generate_frontend_api": {
                    "type": "boolean"
                },
                "generate_frontend_page": {
                    "type": "boolean"
                },
                "generate_frontend_types": {
                    "type": "boolean"
                },
                "generate_model": {
                    "type": "boolean"
                },
                "generate_router": {
                    "type": "boolean"
                },
                "generate_service": {
                    "type": "boolean"
                }
            }
        },
        "tools.TableMetadata": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tools.CodeGenColumnInfo"
                    }
                },
                "table_comment": {
                    "type": "string"
                },
                "table_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT token format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}